Arranging component files into separate folders
