+++ components/CollapsibleTable/CollapsibleTable.tsx
@@ -6,7 +6,7 @@ import TableContainer from "@mui/material/TableContainer";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
 import Paper from "@mui/material/Paper";
-import styles from '../components/CollapsibleTable.module.css'
+import styles from './CollapsibleTable.module.css'
 
 type CollapsibleTableProps = {
 	children: ReactNode	//ment to be a list of rows
