+++ components/DependenciesContainer/DependenciesContainer.module.css
@@ -1,8 +1,5 @@
-/* div.dependenciesStyle {
-  align-items: flex-start;
-  max-width: 200%;
-} */
 
+/* -----------------START: Dependencies CONTAINER CSS ------------------- */
 
 div.sectionContainer {
   display: flex;
