+++ components/DependenciesContainer/DependenciesContainer.tsx
@@ -1,8 +1,8 @@
 import React, { ReactNode } from "react";
-import CollapsibleTable from "./CollapsibleTable";
+import CollapsibleTable from "../CollapsibleTable/CollapsibleTable";
 import styles from "./DependenciesContainer.module.css";
-import SearchBar from "./SearchBar";
-import { DependencyData } from "../src/dataProcessing";
+import SearchBar from "../SearchBar/SearchBar";
+import { DependencyData } from "../../src/dataProcessing";
 // import refreshIcon from "../components/images/refresh.svg" ;
 // import filterIcon from "../components/images/filter.svg" ;
 // import Image from "next/image";
