+++ components/HeaderContainer/HeaderContainer.module.css
@@ -1,0 +1,20 @@
+/* -----------------START: HEADER CONTAINER CSS ------------------- */
+
+div.headerStyle {
+    align-items: center;
+    font-family: "Work Sans", sans-serif;
+  }
+
+  /* H1 Title CSS, overrides the default H1 margins */
+  h2.headerStyle {
+    margin-top: 0;
+    margin-bottom: 0;
+    color: "var(--colour-font)";
+  }
+
+  p.headerStyle {
+    margin-top: 0;
+    margin-bottom: 0;
+    text-align: center;
+  }
+  /* -----------------END: HEADER CONTAINER CSS ------------------- */
