+++ components/Layout/layout.module.css
@@ -8,7 +8,7 @@
 	padding-top: 3.3rem;
 	padding-bottom: 3.3rem;
 	max-height: max-content;
-	
+
   }
 
   .main > section {
