+++ components/Page.tsx
@@ -1,13 +1,13 @@
 import React, { useState } from "react";
-import { InverseSubRow } from "./InverseSubRow";
-import { SubRow } from "./SubRow";
+import { InverseSubRow } from "./SubRow/InverseSubRow";
+import { SubRow } from "./SubRow/SubRow";
 import { useProcessDependencyData } from "../hooks/useProcessDependencyData";
-import Row from "./Row";
-import Layout from "./Layout";
+import Row from "./Row/Row";
+import Layout from "./Layout/Layout";
 import Head from "next/head";
-import DependenciesContainer from "../components/DependenciesContainer";
-import HeaderContainer from "./HeaderContainer";
-import SummaryContainer from "./SummaryContainer";
+import DependenciesContainer from "./DependenciesContainer/DependenciesContainer";
+import HeaderContainer from "./HeaderContainer/HeaderContainer";
+import SummaryContainer from "./SummaryContainer/SummaryContainer";
 import { DependencyData } from "../src/dataProcessing";
 import CircularProgress from "@mui/material/CircularProgress";
 import Box from "@mui/material/Box";
