+++ components/SearchBar/SearchBar.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import Image from "next/image";
-import styles from "../components/treeView.module.css";
-import magnifyingGlass from "./images/magnifying-glass.svg";
+import styles from "../SearchBar/SearchBar.module.css";
+import magnifyingGlass from "../images/magnifying-glass.svg";
 
 
 export default function SearchBar(props: {searchTerm: any, setSearchTerm: any}) {
-    //const [searchTerm, setSearchTerm] = useState<string>('')
 
     return (
             <div className={styles.searchBar}>
