+++ components/SubRow/InverseSubRow.tsx
@@ -1,13 +1,13 @@
 import TableCell from "@mui/material/TableCell";
 import TableRow from "@mui/material/TableRow";
-import { semVerToString } from "../src/semVer";
+import { semVerToString } from "../../src/semVer";
 import { SubRowProps } from "./SubRow";
 import Image from "next/image";
 import styles from "./SubRow.module.css";
 
-import RedIcon from "./images/redIcon.svg";
-import YellowIcon from "./images/yellowIcon.svg";
-import GreenIcon from "./images/greenIcon.svg";
+import RedIcon from "../images/redIcon.svg";
+import YellowIcon from "../images/yellowIcon.svg";
+import GreenIcon from "../images/greenIcon.svg";
 
 type InverseSubRowProps = {
   //This is just renaming dependency to user to make it more clear
