+++ components/treeView.module.css
@@ -1,41 +1,5 @@
-/* Remove default bullets */
-ul.treeView {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-}
 
-/* Style the caret/arrow */
-span.caret {
-  cursor: pointer;
-  user-select: none;
-  /* Prevent text selection */
-}
-
-/* Create the caret/arrow with a unicode, and style it */
-span.caret::before {
-  content: "\25B6";
-  color: black;
-  display: inline-block;
-  margin-right: 6px;
-}
-
-/* Rotate the caret/arrow icon when clicked on (using JavaScript) */
-span.caret_down::before {
-  transform: rotate(90deg);
-}
-
-/* Hide the nested list */
-ul.nested {
-  display: none;
-}
-
-/* Show the nested list when the user clicks on the caret/arrow (with JavaScript) */
-ul.active {
-  display: block;
-}
 /* -----------------START: SHARED SECTION CONTAINERS CSS ------------------- */
-
 div.sectionContainer {
   display: flex;
   flex-direction: column;
@@ -45,14 +9,11 @@ div.sectionContainer {
   height: max-content;
   width: 100%;
   max-width: 2000px;
-
   background-color: var(--colour-container-background);
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
   border-radius: 20px;
   padding: 1.5rem 2rem 1.5rem 2rem;
   margin: 1.1rem;
-  /* margin-top: 1rem;
-  margin-bottom: 1rem; */
   box-sizing: border-box;
   font-family: 'Work Sans', sans-serif;
 }
@@ -61,215 +22,4 @@ h3.h3ContainerStyle {
   margin-bottom: 0.5rem;
   color: "var(--colour-font)";
 }
-
-div.loadingWheelBox{
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-  max-height: 100%;
-  background-color: rgb(242, 240, 238);
-
-}
-
 /* -----------------END: SHARED SECTION CONTAINERS CSS ------------------- */
-
-/* -----------------START: HEADER CONTAINER CSS ------------------- */
-/* Adds the thick black bars around each section, but without a border at the bottom */
-/*
-div.topBarStyle {
-  justify-content: center;
-  height: max-content;
-  width: 85vw;
-  border-bottom: none;
-  padding: 0;
-  margin-top: 2rem;
-  background-color: var(--colour-background);
-}
-*/
-
-/* new ver of div.topBarStyle */
-div.headerStyle {
-  align-items: center;
-  font-family: "Work Sans", sans-serif;
-}
-
-/* H1 Title CSS, overrides the default H1 margins */
-h2.headerStyle {
-  margin-top: 0;
-  margin-bottom: 0;
-  color: "var(--colour-font)";
-}
-
-p.headerStyle {
-  margin-top: 0;
-  margin-bottom: 0;
-  text-align: center;
-}
-/* -----------------END: HEADER CONTAINER CSS ------------------- */
-
-/* -----------------START: SUMMARY CONTAINER CSS ------------------- */
-div.summaryStyle {
-  align-items: flex-start;
-  font-family: "Work Sans", sans-serif;
-  min-width: 15rem;
-}
-
-div.summaryOverall{
-  text-align: center;
-  flex-wrap: wrap;
-  max-height: 100%;
-  background-color: rgb(242, 240, 238);
-  box-sizing: border-box;
-}
-
-div.summaryComponent {
-  /* display: flex;
-  flex-direction: row; */
-  display: block;
-  max-height: 100%;
-  background-color: rgb(242, 240, 238);
-  box-sizing: border-box;
-}
-div.summaryOverall h3{
-  margin-top: 1rem;
-  margin-bottom: 0.5rem;
-}
-
-div.summaryStyle div.container {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  width: 100%;
-  justify-content: space-between;
-  margin-top: 1rem;
-  margin-bottom: 0.5rem;
-
-}
-div.summaryCompHeader h4{
-  margin-top: 0rem;
-  margin-bottom: 0rem;
-}
-
-div.summaryComponent2 {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  width: 100%;
-  background-color: white;
-  border-radius: 20px;
-  box-sizing: border-box;
-  font-size: 10px;
-  padding: 1rem 1rem 1rem 1rem;
-
-}
-
-div.summaryComponent div.summaryCompHeader {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 0.5rem;
-  margin-bottom: 1.5rem;
-  width: 12.5rem;
-}
-
-div.sharedCompProps {
-  width: 31.8%;
-  min-width: 18rem;
-  border-radius: 20px;
-  padding: 1.5rem 2rem 2rem 2rem;
-  box-sizing: border-box;
-}
-
-h2.percentStyle {
-  color: white;
-  font-size: 100px;
-  font-weight: 500;
-  align-self: center;
-}
-
-h3.overallTtitleStyle {
-  color: white;
-  text-align: left;
-}
-
-div.percentComponent {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  align-content: center;
-  justify-content: center;
-}
-
-div.summaryHeader {
-  width: 100%;
-}
-/* -----------------END: SUMMARY CONTAINER CSS ------------------- */
-
-/* -----------------START: Dependencies CONTAINER CSS ------------------- */
-/* Adds the thick black bars around each section */
-div.dependenciesStyle {
-  align-items: flex-start;
-  max-width: 100%;
-}
-
-.depsBarStyle {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  width: 100%;
-
-}
-
-/* -----------------END: SUMMARY CONTAINER CSS ------------------- */
-
-/* -----------------START: SEARCH BAR CONTAINER CSS ------------------- */
-div.searchBar {
-  max-width: 27rem;
-  border: 1px solid #000000;
-  border-radius: 20px;
-  width: 70%;
-  display: flex;
-  flex-direction: row;
-  margin-top: 0.5rem;
-  margin-bottom: 0.5rem;
-  font-family: "Work Sans", sans-serif;
-  /* margin-right: 5rem; */
-  padding-right: 0.75rem;
-}
-
-div.searchBar:focus-within {
-  border-color: rgb(30, 101, 255);
-  border-width: 2px;
-}
-
-/* removes blue outline */
-div.searchBar input[type="text"]:focus {
-  outline: none;
-}
-
-div.searchBar input {
-  padding: 0.65rem;
-  background: #ffffff;
-  border: 0px solid rgba(0, 0, 0, 0);
-  border-radius: 20px;
-
-  width: 100%;
-  height: 100%;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 19px;
-}
-/* -----------------END: SEARCH BAR CONTAINER CSS ------------------- */
-
-.help {
-  float: right;
-  margin-top: 30px;
-  margin-right: 10px;
-}
-
-.helpBtn {
-  cursor: pointer;
-}
